session provider updates
